+++ api/lib/application/usecases/check-user-is-admin-of-certification-center-with-certification-center-invitation-id.js
@@ -1,4 +1,4 @@
-import * as certificationCenterInvitationRepository from '../../infrastructure/repositories/certification-center-invitation-repository.js';
+import * as certificationCenterInvitationRepository from '../../../src/team/infrastructure/repositories/certification-center-invitation-repository.js';
 import * as certificationCenterMembershipRepository from '../../infrastructure/repositories/certification-center-membership-repository.js';
 
 async function execute({
