+++ api/src/team/domain/usecases/index.js
@@ -1,11 +1,12 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-import * as certificationCenterInvitationService from '../../../../lib/domain/services/certification-center-invitation-service.js';
-import * as certificationCenterInvitationRepository from '../../../../lib/infrastructure/repositories/certification-center-invitation-repository.js';
+import * as mailService from '../../../../lib/domain/services/mail-service.js';
 import * as certificationCenterRepository from '../../../certification/shared/infrastructure/repositories/certification-center-repository.js';
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
+import * as certificationCenterInvitationRepository from '../../infrastructure/repositories/certification-center-invitation-repository.js';
+import * as certificationCenterInvitationService from '../services/certification-center-invitation-service.js';
 
 const path = dirname(fileURLToPath(import.meta.url));
 
@@ -13,6 +14,7 @@ const dependencies = {
   certificationCenterRepository,
   certificationCenterInvitationRepository,
   certificationCenterInvitationService,
+  mailService,
 };
 
 const usecasesWithoutInjectedDependencies = {
